Updated openstack/openstack Project: openstack/nova  e740748d9e287e03eee89bc16ccf3060e7a3043a
Enhance notification doc generation with samples

Until now the notification devref only contained a link to the sample file
and that link did only exist after the sample file was merged.
This is suboptimal as the locally generated doc might be incomplete.

This patch changes the doc to contain the content of the sample file.

Change-Id: I4343a2feb13de8b614e8d42d5f1db14102ab492e