Updated openstack/openstack Project: openstack/nova  7cf8ab73544c04ab31429da73cd377b9ea0764eb
Rename compute manager _check_dev_name to _add_missing_dev_names

'check' is a misleading name for this function, as it modifies its
input.

We tweak the loop to be slightly more efficient and readable.

We also save the bdm after updating it. This may have previously been
a bug.

Change-Id: Idadb6192eff68c21cff3387147b98319b6203505