Updated openstack/openstack Project: openstack/heat-templates  7c5aae183aabb3129fc90c60ccdaf5c1f17dd6ba
A hook which invokes os-apply-config.

The intent is for this element (hook script) to be used in place of the
one in tripleo-image-elements which relies on an external signal handling
shell script at the end of the os-refresh-config run
(99-refresh-completed).

This version will run os-apply-config and return a signal immediately.
Because it uses the heat-hook mechanisms it also supports a broader set of
signal handling capabilities... which 99-refresh-completed doesn't fully
support.

Change-Id: Ic9402ff93cbc840bec1debcd8881de563d03cbf0