Updated openstack/openstack Project: openstack/python-keystoneclient  3870b8a763eb473abde37bd44ed3269919408094
Correct test_implied_roles

That test is not supposed to test the count of roles. It is about role
inference rules. This change removes the pure role checks, as it is not
the main intent of the test.

This also fixes a related bug where this test is failing
intermittently. See the bug description for more details.

Change-Id: I8ea1b6d8f344296486427a1f9ebc31c58314501e
Closes-Bug: 1609398