Updated openstack/openstack Project: openstack/ironic  4bd6e9793d5d48ecd90ec1e054549d6754eebae0
Move default network_interface logic in node object

This patch moves the default network_interface logic into the node
object using a field default value. This means if the network_interface
on a node object is ever set to None it'll calculate the default to
place into that field.

Change-Id: I753d969b31ba2d12df4db3fc95a7d9e9ca1e4df6
Partial-Bug: #1608511