Updated openstack/openstack Project: openstack-infra/project-config  5432f3797493e5dd6d42f8d61684436244cb7b00
update tag-release git command to fall back to master

The release tagging script relied on the zuul-cloner behavior of falling
back to master if the requested branch did not exist. Add logic to the
shell command replacement to do the same thing, with a warning.

I tested this locally on the Oslo releases today, and it worked. Before
the change release.sh failed because it could not check out the
stable/newton branch of openstack/requirements to submit the patch to
update the constraints list.

Change-Id: Ib29a68edbb5bfea64625128da35ddad61098dd9c
Signed-off-by: Doug Hellmann <doug@doughellmann.com>