Updated openstack/openstack Project: openstack-infra/puppet-gerrit  53ec1731cc45d1bdaf5109290358b3b72a97f838
Add configurations for gerrit its plugins

This change will allow puppet to configure the its-storyboard plugin.
Settings in gerrit.config and secure.config is to associate gerrit
with an its plugin and settings in actions.config is to tell the
plugin to do something when a gerrit change gets updated. These
conditions are provided by the its-base plugin[1].

[1] https://gerrit.googlesource.com/plugins/its-base/+/stable-2.11/src/main/resources/Documentation/config-common.md
Change-Id: I9f47a2ed88ffbe827e494a478c0dc89a08bbe370