Updated openstack/openstack Project: openstack-infra/shade  732cddbf15b71ae1fe5e7aac58ce406cad286593
Go ahead and admit that we return Munch objects

There was a point in time in the past where we were gung ho on using
Munch as a halfway house to get us to pure dicts. However, we released
1.0 before we could do that, so we're pretty stuck with them for the
forseeable future. That's not terrible - they're pretty low overhead and
work pretty well.

Depends-On: Ie10099430481ffa76f5a19557e3693189544df6b
Change-Id: I8b7dd2c4038db999280ec0c2c9c43fb9499e6d22