Updated openstack/openstack Project: openstack/trove  5dee92154ac23b9a85e21ec633f281c7f399ce8e
Ophaned Volume Not Removed on Instance Delete

When an instance is created with quota issues, the cinder volume is
created butthe nova instance fails to be created. Deleting the
instance would fail todelete the volume. Check the status of the
volume when deleting the instance, if the status of the volume is
"available" delete the volume as it is not connected to any instance.

Change-Id: Ie921a8ff2851e2d9d76a3c3836945c750f090c4e
Closes-Bug: #1344113