Updated openstack/openstack Project: openstack/ironic  761dadba73737c7a237b507c128816a22c632b1a
Fix for "db type could not be determined" error message

If tests are run under py27 then run under py34 or py35, user may get
`db type could not be determined` error.

This patch adds a note to the dev-quickstart guide telling user to
remove the file `.testrepository/times.dbm` and then run the py34 or
py35 test to work around this error.

Change-Id: I57b583eaa8586d14cee730eefc515ad064ac8b32
Closes-Bug: #1229445 (regarding Ironic)