Updated openstack/openstack Project: openstack/tempest  c682bba8984f7dae3c789e52d428edf5dcc1ce58
Optional setting service client on factory

Add the ability to store a copy of an initialised service client
on the service client factory, accessible via an attribute named
as a specified alias.

This allows to stop initialising all clients at __init__ time,
while preserving the ability to initialise them only once, for
instance in the setup_clients phase of tests.

Change-Id: I8b2a907f68214dd1fce1d057408d3ddd1e4bdc0a
Partially-implements: bp client-manager-refactor