Updated openstack/openstack Project: openstack/neutron  f2c481ee3cddd85b73e2fae94af154cfa84b359f
Fix indexerror in delete_csnat_port

The code assumed every port returned from the csnat port query
would have a fixed_ip that it could compare the subnet it is
looking for to. This should be a valid assumption however there
is a path leading to a condition where it has no IPs. This makes
the cleanup code handle this case and dump a warning until we can
figure out what causes the interface to lose the IP.

Partial-Bug: #1609540
Change-Id: Ida024a231bb3fc09dad0e80498f57a5761ca3420