Updated openstack/openstack Project: openstack/zaqar  cf6210df07dc29cb952d33f9d65cec8b49b04718
Don't check for project-id if accessing the / route

The versions route (or / route) should be publicly accessible and is
used for keystone discovery. We shouldn't be checking for the
project-id if that's accessed. On the other hand, this also fixes the
issue where the server crashes if no version is given.

Change-Id: I0481ef6f3920c8ee56ea79ed5621206246b26299
Closes-Bug: #1609662