Updated openstack/openstack Project: openstack/horizon  b0ced2cedc529c4395158c8b9ac442414b4a4fc4
Allow listFunction extra params

---
Includes numerous bug fixes to hz-dynamic-table and friends to be
responsive to dynamic table configurations or trackBy expressions.
---

Some resource types (like subnets) can only be listed for a given
parent container (like networks). This patch modifies the
hz-resource-table to allow a user to pass in extra paramters that
will be passed to the list function for a given resource type.

For example usage: See DNSaaS
https://review.openstack.org/#/c/341182/9/designatedashboard/static/designatedashboard/resources/os-designate-recordset/details/zone-recordsets.html

For example:
This allows hz-resource-table to be used to show a list of subnets
as one of the details views for a network. The details view
controller passes in the current network ID as an extra paramter
to the hz-resource-table, which in turn, supplies that value to
the subnets list function.  The subnets list function can use
the parameters it is given to build the appropriate API call.

This is very similar to what already exists to pass server search
parameters from magic search to the API in the angular images
table.

Change-Id: I90c851aef0a452e4e8ef39e938ac3ca67a93cfac
Partially-Implements: blueprint angular-registry
Closes-Bug: 1608462