Updated openstack/openstack Project: openstack/oslo.vmware  a95530e97daf99161585e7a9190927f9f4518716
Fix image meta-data update for Glance v2

This patch fixes the following issues with Glance v2 image meta-
data update:

* Glance v2 doesn't support updating image size. The size will
be set by the glance service and there is no need to update it.

* We need to convert the image version to string before update
because Glance v2 doesn't support integer values for properties.

This patch also deprecates the image version kwarg because it is
not used by VMware drivers.

Change-Id: I0da5958cbde8be2008ee00602e6e9561bf167bfd
Closes-bug: #1611288