Updated openstack/openstack Project: openstack/heat  1188b4bb3a10bbaaa00a96a0c7062cccd7b0fcf9
Add scenario test_server_signal

This is similar to test_server_cfn_init with the following
differences:
- Use heat instead of AWS resources
- Use HOT template format
- Create and configure a neutron network which the server will be
  booted into

The last difference will allow the tempest test
tempest.api.orchestration.stacks.test_neutron_resources to be deleted.
This tempest test is more appropriate as a scenario test, so it
wouldn't even be considered for defcore and really doesn't belong in
tempest.

Change-Id: Id4f76be2114fb07b3c608d7d0ec0f76f43ad1d21