Updated openstack/openstack Project: openstack/python-glanceclient  07eb350b33aed017a248b40bec07086625271525
Remove unused openstack/common/apiclient/client

This code comes from oslo-incubator and is now frozen,
and the client module specifically is not used in glanceclient.
It's confusing to have in the library since glanceclient has
it's own HTTPClient under glanceclient.common.http. So let's just
remove this effectively dead code.

Change-Id: I820b2439ce7158a63dc55553ce50a9580d63ffb3