Updated openstack/openstack Project: openstack-infra/project-config  470a383e93851fe0949dfcb6c19ce3d92bcb6308
Setup constraints for release

For using constraints in the release pipeline, we need to setup the tree
as follow: The project checked out needs to point to the tag and the
requirements repo needs to point to the corresponding branch. Git does
not easily give us the tag, especially since a tag could be on multiple
branches. Implement some logic that gives us the oldest *stable* branch
the tag is on and use that one. In case no *stable* branch is found, use
master.

The builder can be used in other pipelines as well since it checks
explicitely for tags.

Use this for now only for the {name}-tarball job.

Change-Id: I806062dd24f9dbd2b98cafd32c08bb524e73c5ac