Updated openstack/openstack Project: openstack/neutron  c2b220ee3dbf3583b88980278b1a42a722a74aae
Enable sorting and pagination by default

Now that we have a decent api test coverage for those features, as well
as a way to detect the features via REST API, we should be safe and
thrilled to enable them by default in all installations.

Depends-On: I0aaaa037a8ad52060a68dd75c0a1accc6add238e
Depends-On: I5e68f471a641a34100aba31cb2c4a815c7220014
Depends-On: I648851b48d0481c97054e1280b60a119b42dfd38
Depends-On: Ic04cd125dd715ae0694fbf24d8193abe3151ffd1
Depends-On: Ib3709d15fd87b93285dadf78cce73ee37e8898e2
Depends-On: Ie3a66d2a7c05143df3da1c56c8e8a5308d4bae91

DocImpact Update configuration documentation to cover new defaults.
UpgradeImpact To retain the original default behaviour, set
              allow_sorting and allow_pagination to False.

Closes-Bug: #1566514
Change-Id: I3820bec029b3895913b034b718116c95b4942ed0