Updated openstack/openstack Project: openstack/horizon  7409b1baad2e2022d66c9659507930c578e3410f
Remove old style policy checks

Since the main policy engine moved to django_openstack_auth, the
policy.py file handles the check for POLICY_CHECK_FUNCTION and just
returns True if it is not defined. This eliminates the need for all
the special casing that is being removed in this patch.

One special case (not test related) was left in place, that is in
admin/dashboard.py which has more complex behavior if the operator
is not using policy.

Change-Id: I5523632d3459b68693fcc781bd024aea1180c110