Updated openstack/openstack Project: openstack/neutron  261912d996d8233517d16ae567d358f86f779ff0
Restore old assert_ping behavior

assert_ping() was changed recently to be async-friendly, but
the change caused the drop of a single packet to throw an
error.  Since occasionally the first packet is lost due to
address resolution (ARP) we can't use it for checking liveness
of an IP.

Restored assert_ping() and moved updated code to
assert_async_ping() since that is a special-case.

Change-Id: Ibe69417a0d819d4cd87e2f487c08fd126b1024e2