Updated openstack/openstack Project: openstack/oslo.config  2449197425ea5a712d93191042414c540f2656e4
Advanced Option

Allows for an option to be tagged with advanced true/
false, notifying the end user about uncommon usage.
Such an option is not normally used, and will be
moved to the bottom of the respective namespace.
Only options set to true (default: false) will be
tagged as such in generated sample files.

Change-Id: I2689aba283e473bb293bf0cf0c1c4bcb5979d115
Implements: blueprint advanced-option