Updated openstack/openstack Project: openstack/tripleo-specs  ad2cd11d998a378cf122a73500e750a6a0d32f90
Undercloud Upgrade

Our currently documented upgrade path for the undercloud is very
problematic.  In fact, it doesn't work.  A number of different
patches are attempting to address this problem, but they all have
varying approaches to it that are not necessarily compatible with
each other.  We need to use this spec to decide on the One True
Undercloud Upgrade so we can make progress on the implementation.

Change-Id: I5a8e021336ac688512eef49f154bdd8a21e36929