Updated openstack/openstack Project: openstack-infra/nodepool  3c8bfe5e824371762df3e34d7363cf2e31d59ae6
Allow extra time for initial ZooKeeper connection

We see some random ZK test failures in the pipelines. I believe
this may be because we just aren't waiting long enough for the ZK
process to start up. The kazoo client defaults to waiting only 10
seconds. This change increases the first connection attempt to wait
60 seconds.

Change-Id: I7801166f1a99e1641222e91ef7fc76aad3b6f104