Updated openstack/openstack Project: openstack-infra/tripleo-ci  187e4d30d80be25cc427a160eb8463a7fa86d573
Enable net-iso for ovb jobs

The OVB networking layout differs somewhat from the old CI
environment, so we have to use the templates from OVB.  Ideally we
would use the templates directly from the OVB repo, but it isn't
part of OpenStack Git yet, so for the moment we just use a copy of
them.  They don't/can't change much without breaking existing users
anyway, so this is low risk copy pasta.

Change-Id: I0a2c8ca49cda2536b25884fcfeb3720b47c47863