Updated openstack/openstack Project: openstack/cinder  c5d50362ee74e8de30a893891a340b8c2b918430
Improve snapshot handling in K2

In creating snapshot, we use "Best Effort" retention policy.
In this policy snapshot may be deleted in case capacity is running
out, causing an inconsistency between K2 state and cinder DB.

In order to avoid this inconsistency, add "is_auto_deleteable=False"
to the snapshot properties, in order to avoid auto deleting
the snapshot.
Closes-Bug: #1608488
Change-Id: I70f770ce42ff56314b893a48ff7ed1a46a4ae55f