Updated openstack/openstack Project: openstack-infra/system-config  c622221dc768e6c676834c4d67f25bc1dd2c3625
Set the gitpgpkey for signing nodes

Since our infrastructure signing key includes a unique comment field
in its name which changes with the key each cycle when it's rotated,
we need to explicitly let Git know how to identify the key (it
doesn't automatically rely on the default secret key in the keychain
and instead expects that any optional comment be part of the name).
Use an E-mail address to identify the signing key to Git since this
will remain the same through rotation, and since the
jenkins::jenkinsuser module completely replaces the keyring files
we can simply insure our process only ever sets one key for that
address in the provided keyring contents (which is how we document
doing it already).

Change-Id: I78fe7fa331ae9b51ff0fbf91c0c2f8961b09eb88
Depends-On: Id577c4479abd2f443f73c630a12e94d3ebe63660