Updated openstack/openstack Project: openstack-infra/git-review  cf4e7cdb5fbbfcf0b828e35ad7df69f97010d441
Clarify that submitting multiple commits is OK

I've heard feedback that the message that appears when you send multiple
commits for the review appears to imply that Gerrit cannot handle
multiple patches in a series, and that users should avoid submitting
them in favour of squashing all of their patches before submission. Of
course nothing could be further from the truth, and Gerrit's ability to
gracefully handle patch series is one of its major advantages over other
code review tools.

This changes the message to clarify that multiple self-contained patches
are perfectly OK.

Change-Id: Ia659149ce9332909b4bdbf09a2817a2015ebafd2