Updated openstack/openstack Project: openstack/nova  b3694f75608f17ca7c4faaa1c09143e629466b43
PCI: Fix network calls order on finish_revert_resize()

The driver was spawning the guest before migrate_instance_finish() was being
called which caused the neutron ports to not be updated with the old PCI
devices information.  Doing so results in spawning the guest to go in error
since the driver is using the wrong PCI devices.

Change-Id: I9b715f638fda65dd2be6c6693e55c8502156ed57
Partial-bug: #1512880