Updated openstack/openstack Project: openstack/cinder  b67d13ef4b9ac7d0bba92250ac5a77c15d8e9e68
SwiftBackupDriver: Dont add project_id to authurl

If using the config parameter backup_swift_auth_url for the
Swift backup driver in single_user mode, the project_id is
added to the auth_url. As the auth_url is expected to be a
keystone endpoint, this is not needed. This patch will fix
this.

Change-Id: I865fb2f1dff188c2886ee0ea35a8f0f3c48724b8
Closes-Bug: #1518199